+++ tests/test_dataset.py
@@ -1048,20 +1048,20 @@ def test_writing_subset_of_compressed_data() -> None:
     # create uncompressed dataset
     Dataset.create(TESTOUTPUT_DIR / "compressed_data", scale=(1, 1, 1)).add_layer(
         "color", LayerCategories.COLOR_TYPE
-    ).add_mag("1", block_len=8, file_len=8).write(
-        (np.random.rand(20, 40, 60) * 255).astype(np.uint8)
+    ).add_mag("2", block_len=8, file_len=8).write(
+        (np.random.rand(120, 140, 160) * 255).astype(np.uint8)
     )
 
     # compress data
     compress_mag_inplace(
         (TESTOUTPUT_DIR / "compressed_data").resolve(),
         layer_name="color",
-        mag=Mag("1"),
+        mag=Mag("2"),
     )
 
     # open compressed dataset
     compressed_mag = (
-        Dataset(TESTOUTPUT_DIR / "compressed_data").get_layer("color").get_mag("1")
+        Dataset(TESTOUTPUT_DIR / "compressed_data").get_layer("color").get_mag("2")
     )
 
     with warnings.catch_warnings():
@@ -1081,6 +1081,26 @@ def test_writing_subset_of_compressed_data() -> None:
                 offset=(10, 20, 30),
                 data=(np.random.rand(10, 10, 10) * 255).astype(np.uint8),
             )
+
+        assert compressed_mag._mag_view_bounding_box_at_creation == BoundingBox(
+            topleft=(
+                0,
+                0,
+                0,
+            ),
+            size=(120, 140, 160),
+        )
+        # Writing unaligned data to the edge of the bounding box of the MagView does not raise an error.
+        # This write operation writes unaligned data into the bottom-right corner of the MagView.
+        compressed_mag.write(
+            offset=(64, 64, 64),
+            data=(np.random.rand(56, 76, 96) * 255).astype(np.uint8),
+        )
+        # This also works for normal Views but they only use the bounding box at the time of creation as reference.
+        compressed_mag.get_view().write(
+            offset=(64, 64, 64),
+            data=(np.random.rand(56, 76, 96) * 255).astype(np.uint8),
+        )
 
         # Writing aligned data does not raise a warning. Therefore, this does not fail with these strict settings.
         compressed_mag.write(data=(np.random.rand(64, 64, 64) * 255).astype(np.uint8))