+++ tests/test_downsampling.py
@@ -1,4 +1,5 @@
 import logging
+import warnings
 from pathlib import Path
 from typing import Tuple, cast
 
@@ -21,7 +22,6 @@ from wkcuber.utils import WkwDatasetInfo, open_wkw
 from wkcuber.downsampling_utils import _mode, non_linear_filter_3d
 import shutil
 
-WKW_CUBE_SIZE = 1024
 CUBE_EDGE_LEN = 256
 
 TESTOUTPUT_DIR = Path("testoutput")
@@ -93,23 +93,20 @@ def downsample_test_helper(use_compress: bool) -> None:
         pass
 
     source_ds = Dataset(source_path)
-    source_layer = source_ds.get_layer("color")
-    mag1 = source_layer.get_mag("1")
+    target_ds = source_ds.copy_dataset(target_path, block_len=16, file_len=16)
 
-    target_ds = Dataset.create(target_path, scale=(1, 1, 1))
-    target_layer = target_ds.add_layer(
-        "color", LayerCategories.COLOR_TYPE, dtype_per_channel="uint8"
-    )
-    # The bounding box has to be set here explicitly because the downsampled data is written to a different dataset.
-    target_layer.set_bounding_box(
-        source_layer.get_bounding_box().topleft, source_layer.get_bounding_box().size
-    )
+    target_layer = target_ds.get_layer("color")
+    mag1 = target_layer.get_mag("1")
+    target_layer.delete_mag("2-2-1")  # This is not needed for this test
+
     mag2 = target_layer._initialize_mag_from_other_mag("2", mag1, use_compress)
 
-    offset = (WKW_CUBE_SIZE, 2 * WKW_CUBE_SIZE, 0)
-    target_offset = cast(Tuple[int, int, int], tuple([o // 2 for o in offset]))
-    source_size = cast(Tuple[int, int, int], (CUBE_EDGE_LEN * 2,) * 3)
-    target_size = cast(Tuple[int, int, int], (CUBE_EDGE_LEN,) * 3)
+    # The actual size of mag1 is (4600, 4600, 512).
+    # To keep this test case fast, we are only downsampling a small part
+    offset = (4096, 4096, 0)
+    target_offset = (2048, 2048, 0)
+    source_size = (504, 504, 512)
+    target_size = (252, 252, 256)
     source_buffer = mag1.read(
         offset=offset,
         size=source_size,
@@ -127,7 +124,7 @@ def downsample_test_helper(use_compress: bool) -> None:
         ),
         [2, 2, 2],
         InterpolationModes.MAX,
-        CUBE_EDGE_LEN,
+        128,
         100,
     )
 
@@ -147,7 +144,9 @@ def test_downsample_cube_job() -> None:
 
 
 def test_compressed_downsample_cube_job() -> None:
-    downsample_test_helper(True)
+    with warnings.catch_warnings():
+        warnings.filterwarnings("error")  # This escalates the warning to an error
+        downsample_test_helper(True)
 
 
 def test_downsample_multi_channel() -> None: